fix(FiveHourForecast): fix IndexOutOfBoundsException
- Fixed a problem where an IndexOutOfBoundsException was thrown when
populating the UI for the five day forecast overview in the
DayWeatherFragment.
The error was thrown as a result of hardcoding the indexing in order to
retrieve the hourly data. More specifically,
the same day ( weekDayForecast[0].listOfHourlyData[i] ) was used to get
the hourly data (looped 5 times to get the five day forecast), no backup
was implemented if hoursUntilMidnight < 5