refactor(fragments): change lifecycle handling
Reason for change:
- The view hieracy can be replaced during runtime and
can be re-created multiple times. Any reference that is not initalized
in onCreateView will either be null or point to an old view-hieracy that
has been replaced by a new one and might cause memory leak.
- Also by having a similar lifecycle to activity it becomes easier to reduce the
complexity of the lifecycles, instead of learning two approaches I can
just learn one.